fix(Config_parese): fix param type of token_compare_hash

+++ src/config_parsing/analyse.h
@@ -129,7 +129,7 @@ int ganesha_yyparse(struct parser_state *st);
 int ganeshun_yy_init_parser(char *srcfile, struct parser_state *st);
 void ganeshun_yy_cleanup_parser(struct parser_state *st);
 
-static inline int token_compare_hash(const u_int64_t key1, const u_int16_t key2,
+static inline int token_compare_hash(const u_int64_t key1, const u_int64_t key2,
 				     const char *token1, const char *token2)
 {
 	if (key1 < key2)
